refactor: Column ComponentをColumnsAreaに変更
