+++ src/components/Templates/ColumnsArea/styles.ts
